requirements.txt
eyed3
hsaudiotag
progressbar
path.py
mutagen
docopt
ansitagcolor